Setup on pc

+++ setup/sym.sh
@@ -1,7 +1,22 @@
 #!/usr/bin/env bash
+if [ $# -ne 1 ]; then
+  echo "Usage: $0 <hostname"
+  exit 1
+fi
+
+
+hostname="$1"
+
+cd ~/nixos-config/hosts
+
+mkdir $hostname
 
 sudo rm /etc/nixos/configuration.nix
 
-sudo ln -s ~nixos-config/configuration.nix /etc/nixos/configuration.nix
+sudo ln -s ~/nixos-config/$hostname/configuration.nix /etc/nixos/configuration.nix
+
+cp /etc/nixos/hardware-configuration.nix ~/nixos-config/hosts/$hostname/
+
+sudo rm /etc/nixos/hardware-configuration.nix
 
 echo "Configuration.nix symlink created, remember to do something with hardware-configuration"
